gumroad_product/quant_trading_bot_starter/requirements.txt
# Core Data & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Technical Analysis
ta-lib>=0.4.25
pandas-ta>=0.3.14b

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Data Sources
yfinance>=0.2.30
alpha-vantage>=2.3.1
ccxt>=4.0.0  # Crypto exchanges

# Live Trading
alpaca-trade-api>=3.0.0
ib_insync>=0.9.70  # Interactive Brokers

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0

# Optional: Deep Learning (uncomment if needed)
# torch>=2.0.0
# tensorflow>=2.13.0
